chore: Move check_docker
