+++ opthub_runner_admin/main.py
@@ -8,12 +8,11 @@ from types import FrameType
 from typing import TYPE_CHECKING, Any, cast
 
 import click
-import docker
 import yaml
 from botocore.exceptions import ClientError
-from docker.errors import DockerException
 
 from opthub_runner_admin.utils.credentials.credentials import Credentials
+from opthub_runner_admin.utils.docker import check_docker
 
 if TYPE_CHECKING:
     from opthub_runner_admin.args import Args
@@ -73,21 +72,6 @@ def auth(username: str, password: str) -> None:
         sys.exit(1)
     except Exception as e:
         # another exception
-        click.echo(f"An unexpected error occurred: {e}")
-        sys.exit(1)
-
-
-def check_docker() -> None:
-    """Check if Docker is running and accessible."""
-    try:
-        client = docker.from_env()
-        client.ping()
-    except DockerException as error:
-        click.echo(
-            f"Error: Unable to communicate with Docker. Please ensure Docker is running and accessible. ({error!s})",
-        )
-        sys.exit(1)
-    except Exception as e:
         click.echo(f"An unexpected error occurred: {e}")
         sys.exit(1)
